Soil Humidity Card component has been split into sub components.
